requirements.txt
python-dotenv
pipecat-ai[openai,websocket]=0.0.42
fastapi
uvicorn
loguru
